# node: serve JSON-RPC on all unused HTTP paths