--- a/node/rpcstack.go
+++ b/node/rpcstack.go
@@ -171,6 +171,19 @@ func (h *httpServer) start() error {
 
 func (h *httpServer) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	rpc := h.httpHandler.Load().(*rpcHandler)
+	if rpc != nil {
+		// try and route in the mux first
+
+		// Requests to a path below root are handled by the mux,
+		// which has all the handlers registered via Node.RegisterHandler.
+		// These are made available when RPC is enabled.
+		muxHandler, pattern := h.mux.Handler(r)
+		if pattern != "" {
+			muxHandler.ServeHTTP(w, r)
+			return
+		}
+	}
+	// fall back to JSON-RPC
 	if r.RequestURI == "/" {
 		// Serve JSON-RPC on the root path.
 		ws := h.wsHandler.Load().(*rpcHandler)
@@ -182,12 +195,6 @@ func (h *httpServer) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 			rpc.ServeHTTP(w, r)
 			return
 		}
-	} else if rpc != nil {
-		// Requests to a path below root are handled by the mux,
-		// which has all the handlers registered via Node.RegisterHandler.
-		// These are made available when RPC is enabled.
-		h.mux.ServeHTTP(w, r)
-		return
 	}
 	w.WriteHeader(404)
 }