@ryanschneider 2020-11-13
I ran a mix of random state-related RPCs from a sample of production traffic against our slightly forked version of geth both without and with this patch applied, and actually your PR ran faster than the old one:

Baseline:

```
Summary:
  Total:	30.0628 secs
  Slowest:	1.5710 secs
  Fastest:	0.0004 secs
  Average:	0.0173 secs
  Requests/sec:	2887.2512

Latency distribution:
  10% in 0.0032 secs
  25% in 0.0060 secs
  50% in 0.0102 secs
  75% in 0.0198 secs
  90% in 0.0352 secs
  95% in 0.0479 secs
  99% in 0.1243 secs
```

With this PR merged in:

```
Summary:
  Total:	30.0280 secs
  Slowest:	0.6459 secs
  Fastest:	0.0004 secs
  Average:	0.0137 secs
  Requests/sec:	3649.6590

Latency distribution:
  10% in 0.0025 secs
  25% in 0.0050 secs
  50% in 0.0090 secs
  75% in 0.0165 secs
  90% in 0.0272 secs
  95% in 0.0377 secs
  99% in 0.0820 secs
```

Anyways I think the key takeaway is that this definitely does _not_ have any appreciable negative impact on RPC performance.  I suspect the speed boost was due to the short nature of the test and the distinct possibility that the node was processing more complex blocks or devp2p traffic or what not during the first run.

So in short great work @renaynay and lgtm!  Thanks @holiman for letting me know about this PR, it'll be one less patch for us to keep in our fork going forward.