@fjl 2020-11-17
We discussed this a bit last couple days, and I'm not super happy to include this change. It does work, but the behavior also gets a bit unpredictable. For example, if you enable GraphQL, JSON-RPC will not be served on `/graphql` and RPC requests sent to that path will trigger an error. But without GraphQL, JSON-RPC requests against that path will work, and GraphQL requests sent to that will return a JSON-RPC error.

I think it would be better to restrict JSON-RPC to a single path. @ryanschneider AFAIK, your issue with this was mostly that AWS load balancing does not support changing the path. Can you give us some insight where the path is coming from in your case? I'm hoping it can be solved in a different way.