@ryanschneider 2020-11-17
Without going too deep into specifics, we have multiple groups of nodes behind a single load balancer.  While this isn't exactly how we are setup, imagine that some of our mainnet archive nodes are on a URL like `http://internal-loadbalancer/mainnet/archive`.  Since AWS ALBs don't support rewriting the request, the request to the node is still `POST /mainnet/archive`.  This worked up to 1.9.12 but required us to patch versions .13 and higher.

It could be worked around by running a separate local reverse proxy on the node (e.g. nginx with a rewrite rule) but that seems like a lot of extra devops surface when we already have a proxy in front of the internal ALB in our stack.  Or we could work around the ALB limitation by using Host headers or DNS names instead of URL paths.  The issue isn't really that there aren't other ways to do it, it's that having geth behave like it did before .12 is the least significant change, though I do understand your concern about unexpected behavior.

What about a compromise where the "any path" behavior only works if `h.mux` is "empty" (that is no calls to .RegisterHandler have been made)?  That would behave just like .12 when graphQL is disabled, and trigger the new behavior only once graphQL or some other new handler is added.  And, as mentioned, the PR fixes the odd behavior that `POST /?foo=bar` returns a 404 as well (because of the overly strict `.RequestURI` check vs. parsing the `URL.Path`).