@ryanschneider 2020-11-20
@fjl or @renaynay any thoughts on this compromise approach https://github.com/ethereum/go-ethereum/pull/21826#issuecomment-729058764 ?

> What about a compromise where the "any path" behavior only works if h.mux is "empty" (that is no calls to .RegisterHandler have been made)? That would behave just like .12 when graphQL is disabled, and trigger the new behavior only once graphQL or some other new handler is added. And, as mentioned, the PR fixes the odd behavior that POST /?foo=bar returns a 404 as well (because of the overly strict .RequestURI check vs. parsing the URL.Path).