@fjl 2020-12-17
I have a better idea. Instead of serving JSON-RPC requests on all unmatched paths, we can add a flag that lets you set the path (or path prefix, if necessary). @ryanschneider this would solve your case, since the path is specific to each geth instance, right?