@ryanschneider 2020-12-17
@fjl ya that would work. We would probably run with a wildcard prefix like `/` since we tend to want to move geth nodes around without having to reconfigure them so as long as that was a supported prefix this idea would work for us.