@renaynay 2020-12-18
Okay sounds good. On holiday now but will implement this ASAP.