@ryanschneider 2020-12-18
Awesome, so I'm not 100% familiar with this code but I _think_ that this approach can actually lead to a simpler codebase, since the cleanest thing to do would probably be "register" the Graph QL, GraphQL UI, HTTP JSONRPC and Websocket JSONRPC handlers all the same way, which could allow `httpHandler.ServeHTTP` to just be a mux lookup (though probably still with a special case for `isWebsocket`).