@renaynay 2021-01-12
Option 1: https://github.com/renaynay/go-ethereum/pull/31/files
Option 2: (replace mux router with gorilla mux) https://github.com/renaynay/go-ethereum/pull/32

@fjl